Update README.md
A line following robot is a wheeled robot which follows a track of colored path. Usually a black line is the track . It uses its IR sensors to detect the black marks and convert it into pwm signal. The PWM signal is then conveyed to the dc motors which drive the robot. This particular one which  we've built has 2 dc motors, a motor driver , a 9V Lipo battery and uses the arduino  nano development board as micro controller.